requirements.txt
fastapi
uvicorn
joblib
scikit-learn
python-dotenv
huggingface-hub
